+++ website/theme-builder/components/icons/index.js
