--- website/theme-creator/components/logo/index.js
@@ -1,54 +1,0 @@
-import styles from "./Logo.module.scss";
-
-export default function Logo({ ...props }) {
-  return (
-    <svg
-      className={styles.logo}
-      width="1035"
-      height="98"
-      viewBox="0 0 1035 98"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M38 8H8V75H38V8ZM0 0V83H46V0H0Z"
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M76 23H46V90H76V23ZM38 15V98H84V15H38Z"
-      />
-      <g>
-        <path d="M164.081 58.5268C164.081 50.2072 159.109 45.0169 149.166 42.7271L141.517 40.8953C133.716 39.0634 131.039 37.3843 131.039 33.7969C131.039 29.2936 135.399 26.6985 141.441 26.6985C147.636 26.6985 152.914 29.7516 153.602 36.0104H162.551C161.633 26.9275 154.75 19.6001 141.288 19.6001C130.656 19.6001 121.937 25.5536 121.937 34.4075C121.937 41.7349 127.062 46.5435 137.387 48.9859L145.036 50.7415C151.843 52.3443 155.056 54.6341 155.056 58.9084C155.056 63.7934 150.084 66.6938 143.353 66.6938C135.934 66.6938 130.351 61.3509 129.815 54.4815H120.866C121.708 64.0223 129.28 73.7922 143.277 73.7922C155.514 73.7922 164.081 67.3044 164.081 58.5268Z" />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M194.341 65.0146H194.57V73.0289H202.754V34.8655H194.57V42.8035H194.341C192.505 37.5369 188.069 34.1022 182.256 34.1022C172.083 34.1022 166.194 43.2614 166.194 53.9472C166.194 64.6329 171.854 73.7922 182.256 73.7922C188.069 73.7922 192.505 70.2811 194.341 65.0146ZM174.99 53.9472C174.99 45.7802 178.661 40.2847 184.856 40.2847C191.052 40.2847 194.876 45.7802 194.876 53.9472C194.876 62.0378 191.052 67.6097 184.856 67.6097C178.738 67.6097 174.99 62.0378 174.99 53.9472Z"
-        />
-        <path d="M217.513 50.1308V73.0289H209.176V34.8655H217.284V42.3455H217.437C219.043 37.0026 223.02 34.1022 228.604 34.1022C237.017 34.1022 241.147 39.8267 241.147 48.2227V73.0289H232.81V50.6651C232.81 43.5667 230.286 40.361 225.162 40.361C220.037 40.361 217.513 43.872 217.513 50.1308Z" />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M273.337 65.0146H273.566V73.0289H281.75V20.3634H273.566V42.8035H273.337C271.501 37.5369 267.065 34.1022 261.252 34.1022C251.079 34.1022 245.19 43.2614 245.19 53.9472C245.19 64.6329 250.85 73.7922 261.252 73.7922C267.065 73.7922 271.501 70.2811 273.337 65.0146ZM253.986 53.9472C253.986 45.7802 257.581 40.2847 263.929 40.2847C270.354 40.2847 273.872 45.7802 273.872 53.9472C273.872 62.0378 270.354 67.6097 263.929 67.6097C257.581 67.6097 253.986 62.0378 253.986 53.9472Z"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M296.356 65.0146V87.611H288.172V34.8655H296.356V42.8035H296.585C298.421 37.5369 302.857 34.1022 308.67 34.1022C318.843 34.1022 324.732 43.2614 324.732 53.9472C324.732 64.6329 319.072 73.7922 308.67 73.7922C302.857 73.7922 298.421 70.2811 296.585 65.0146H296.356ZM305.993 40.2847C299.568 40.2847 296.05 45.7802 296.05 53.9472C296.05 62.0378 299.568 67.6097 305.993 67.6097C312.342 67.6097 315.937 62.0378 315.937 53.9472C315.937 45.7802 312.342 40.2847 305.993 40.2847Z"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M355.246 65.0146H355.475V73.0289H363.659V34.8655H355.475V42.8035H355.246C353.41 37.5369 348.974 34.1022 343.161 34.1022C332.988 34.1022 327.099 43.2614 327.099 53.9472C327.099 64.6329 332.759 73.7922 343.161 73.7922C348.974 73.7922 353.41 70.2811 355.246 65.0146ZM335.895 53.9472C335.895 45.7802 339.566 40.2847 345.761 40.2847C351.957 40.2847 355.781 45.7802 355.781 53.9472C355.781 62.0378 351.957 67.6097 345.761 67.6097C339.642 67.6097 335.895 62.0378 335.895 53.9472Z"
-        />
-        <path d="M395.092 59.748H403.582C402.129 68.7546 395.321 73.7922 386.373 73.7922C375.129 73.7922 367.939 64.4803 367.939 53.9472C367.939 43.6431 374.9 34.1022 386.067 34.1022C395.015 34.1022 401.746 38.9871 403.199 47.9937H394.786C393.868 43.1088 390.732 40.2847 386.22 40.2847C380.024 40.2847 376.735 46.0092 376.735 53.9472C376.735 61.9615 380.254 67.6097 386.602 67.6097C391.038 67.6097 394.251 64.6329 395.092 59.748Z" />
-        <path d="M431.968 73.0289H442.447L424.932 51.6574L441.682 34.8655H431.357L415.6 50.5888V20.3634H407.263V73.0289H415.6V53.1076L431.968 73.0289Z" />
-      </g>
-      <path d="M531.288 38.9423C529.486 28.9838 521.52 23.1036 511.751 23.1036C499.326 23.1036 490.316 32.6827 490.316 48.0472C490.316 63.4117 499.326 72.9908 511.751 72.9908C521.52 72.9908 529.486 67.1106 531.288 57.1521H525.408C523.986 63.8859 518.2 67.4899 511.751 67.4899C502.93 67.4899 496.007 60.6613 496.007 48.0472C496.007 35.4331 502.93 28.6044 511.751 28.6044C518.2 28.6044 523.986 32.2085 525.408 38.9423H531.288ZM559.645 57.4366C559.645 64.2653 554.429 67.3951 550.256 67.3951C545.608 67.3951 542.289 63.9808 542.289 58.6696V35.9073H536.693V59.0489C536.693 68.3435 541.625 72.8011 548.454 72.8011C553.955 72.8011 557.559 69.861 559.266 66.1621H559.645V72.3269H565.241V35.9073H559.645V57.4366ZM597.949 44.0638C596.194 38.8949 592.258 35.4331 585.05 35.4331C577.368 35.4331 571.677 39.7959 571.677 45.9606C571.677 50.9873 574.665 54.3542 581.351 55.9191L587.421 57.3418C591.096 58.1953 592.827 59.9499 592.827 62.4633C592.827 65.5931 589.508 68.1538 584.291 68.1538C579.715 68.1538 576.846 66.1858 575.85 62.2736L570.539 63.6014C571.843 69.7899 576.941 73.0856 584.386 73.0856C592.851 73.0856 598.612 68.4621 598.612 62.1787C598.612 57.1047 595.435 53.9037 588.939 52.3151L583.532 50.9873C579.217 49.9203 577.273 48.474 577.273 45.6761C577.273 42.5463 580.592 40.2701 585.05 40.2701C589.934 40.2701 591.95 42.9731 592.922 45.4864L597.949 44.0638ZM619.536 35.9073H611.759V27.1818H606.163V35.9073H600.662V40.6495H606.163V63.4117C606.163 69.7662 611.285 72.8011 616.027 72.8011C618.113 72.8011 619.441 72.4217 620.2 72.1372L619.062 67.1106C618.588 67.2054 617.829 67.3951 616.596 67.3951C614.13 67.3951 611.759 66.6363 611.759 61.8942V40.6495H619.536V35.9073ZM639.019 73.0856C648.883 73.0856 655.522 65.5931 655.522 54.3068C655.522 42.9257 648.883 35.4331 639.019 35.4331C629.155 35.4331 622.516 42.9257 622.516 54.3068C622.516 65.5931 629.155 73.0856 639.019 73.0856ZM639.019 68.059C631.526 68.059 628.112 61.6097 628.112 54.3068C628.112 47.0039 631.526 40.4598 639.019 40.4598C646.512 40.4598 649.926 47.0039 649.926 54.3068C649.926 61.6097 646.512 68.059 639.019 68.059ZM660.725 72.3269H666.321V49.5647C666.321 44.2298 670.209 40.4598 674.572 40.4598C678.816 40.4598 681.78 43.2339 681.78 47.3833V72.3269H687.471V48.6162C687.471 43.9215 690.411 40.4598 695.532 40.4598C699.516 40.4598 702.93 42.57 702.93 47.9523V72.3269H708.526V47.9523C708.526 39.3928 703.926 35.4331 697.429 35.4331C692.213 35.4331 688.395 37.8279 686.522 41.5979H686.143C684.341 37.7093 681.164 35.4331 676.374 35.4331C671.632 35.4331 668.123 37.7093 666.605 41.5979H666.131V35.9073H660.725V72.3269ZM715.418 72.3269H721.014V35.9073H715.418V72.3269ZM718.264 29.8374C720.445 29.8374 722.247 28.1302 722.247 26.0437C722.247 23.9572 720.445 22.25 718.264 22.25C716.082 22.25 714.28 23.9572 714.28 26.0437C714.28 28.1302 716.082 29.8374 718.264 29.8374ZM726.881 72.3269H755.429V67.1106H734.658V66.7312L754.765 40.3649V35.9073H727.545V41.1237H747.557V41.503L726.881 68.059V72.3269ZM771.426 73.1805C777.781 73.1805 781.1 69.7662 782.239 67.3951H782.523V72.3269H788.119V48.3317C788.119 36.7609 779.298 35.4331 774.651 35.4331C769.15 35.4331 762.891 37.33 760.045 43.9689L765.357 45.8658C766.589 43.2102 769.506 40.3649 774.841 40.3649C779.986 40.3649 782.523 43.0917 782.523 47.7627V47.9523C782.523 50.6554 779.773 50.4183 773.134 51.2718C766.376 52.1491 759.002 53.6429 759.002 61.9891C759.002 69.1023 764.503 73.1805 771.426 73.1805ZM772.28 68.1538C767.822 68.1538 764.598 66.1621 764.598 62.2736C764.598 58.0057 768.486 56.6779 772.849 56.1088C775.22 55.8243 781.575 55.1604 782.523 54.0223V59.1438C782.523 63.6962 778.919 68.1538 772.28 68.1538ZM810.832 35.9073H803.055V27.1818H797.46V35.9073H791.959V40.6495H797.46V63.4117C797.46 69.7662 802.581 72.8011 807.323 72.8011C809.41 72.8011 810.738 72.4217 811.496 72.1372L810.358 67.1106C809.884 67.2054 809.125 67.3951 807.892 67.3951C805.426 67.3951 803.055 66.6363 803.055 61.8942V40.6495H810.832V35.9073ZM815.911 72.3269H821.507V35.9073H815.911V72.3269ZM818.756 29.8374C820.938 29.8374 822.74 28.1302 822.74 26.0437C822.74 23.9572 820.938 22.25 818.756 22.25C816.575 22.25 814.773 23.9572 814.773 26.0437C814.773 28.1302 816.575 29.8374 818.756 29.8374ZM843.213 73.0856C853.076 73.0856 859.715 65.5931 859.715 54.3068C859.715 42.9257 853.076 35.4331 843.213 35.4331C833.349 35.4331 826.71 42.9257 826.71 54.3068C826.71 65.5931 833.349 73.0856 843.213 73.0856ZM843.213 68.059C835.72 68.059 832.306 61.6097 832.306 54.3068C832.306 47.0039 835.72 40.4598 843.213 40.4598C850.705 40.4598 854.12 47.0039 854.12 54.3068C854.12 61.6097 850.705 68.059 843.213 68.059ZM870.514 50.4183C870.514 44.0638 874.45 40.4598 879.809 40.4598C885.002 40.4598 888.155 43.8504 888.155 49.5647V72.3269H893.751V49.1853C893.751 39.8907 888.795 35.4331 881.421 35.4331C875.921 35.4331 872.506 37.899 870.799 41.5979H870.325V35.9073H864.919V72.3269H870.514V50.4183ZM914.181 28.9838H929.451V72.3269H935.331V28.9838H950.601V23.7675H914.181V28.9838ZM965.093 73.0856C974.956 73.0856 981.595 65.5931 981.595 54.3068C981.595 42.9257 974.956 35.4331 965.093 35.4331C955.229 35.4331 948.59 42.9257 948.59 54.3068C948.59 65.5931 955.229 73.0856 965.093 73.0856ZM965.093 68.059C957.6 68.059 954.186 61.6097 954.186 54.3068C954.186 47.0039 957.6 40.4598 965.093 40.4598C972.585 40.4598 976 47.0039 976 54.3068C976 61.6097 972.585 68.059 965.093 68.059ZM1001.59 73.0856C1011.46 73.0856 1018.1 65.5931 1018.1 54.3068C1018.1 42.9257 1011.46 35.4331 1001.59 35.4331C991.731 35.4331 985.092 42.9257 985.092 54.3068C985.092 65.5931 991.731 73.0856 1001.59 73.0856ZM1001.59 68.059C994.102 68.059 990.687 61.6097 990.687 54.3068C990.687 47.0039 994.102 40.4598 1001.59 40.4598C1009.09 40.4598 1012.5 47.0039 1012.5 54.3068C1012.5 61.6097 1009.09 68.059 1001.59 68.059ZM1028.9 23.7675H1023.3V72.3269H1028.9V23.7675Z" />
-      <g>
-        <path d="M476.201 20.0105H471.467L457.002 73.7504H461.736L476.201 20.0105Z" />
-      </g>
-    </svg>
-  );
-}
